If this works I'll be very angry...
This is my attempt at implimenting the gskinner swfbridge into Community 5b for some fun stuff (like the custom mouse and mostly stuff that only applies to me). I tried for hours to do it yesterday and I just had a brainstorm so I'll be pretty steamed if this works. Also overjoyed so I should probably change the title of this.